MLServer/requirements.txt
boto==2.42.0
bz2file==0.98
geopy==1.11.0
httpretty==0.8.14
mysql-connector-python-rf==2.2.1
numpy==1.11.1
requests==2.11.1
scikit-learn==0.18
scipy==0.18.1
smart-open==1.3.2
yolk==0.4.3
